brooklyn-software-network: add org.apache package prefix
