+++ software/network/src/main/java/org/apache/brooklyn/entity/network/bind/BindOsSupport.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package brooklyn.entity.network.bind;
+package org.apache.brooklyn.entity.network.bind;
 
 import javax.annotation.concurrent.Immutable;
