+++ software/network/src/main/resources/org/apache/brooklyn/entity/network/bind/named.loopback
