+++ software/network/src/test/java/org/apache/brooklyn/entity/network/bind/TestBindDnsServerImpl.java
@@ -16,12 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package brooklyn.entity.network.bind;
+package org.apache.brooklyn.entity.network.bind;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.google.common.annotations.VisibleForTesting;
 
 import brooklyn.entity.Entity;
 import brooklyn.entity.basic.EntityLocal;
